Refactor code by moving visualization() and print_performance() from motion2.py to visualization.py
